refactor: LottoMaker 클래스 제거
